requirements.txt
chainlit>=0.7.700
fastapi>=0.103.1
httpx>=0.25.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.3.0
openai>=1.3.0
langchain>=0.0.267
langchain-openai>=0.0.2